chore: refactor comments and imports
