+++ the-easy-score/src/components/Loading.js
@@ -1,5 +1,6 @@
 import React from "react";
 import { Spin } from "antd";
+
 import "../App.css";
 
 const Loading = () => {
