+++ the-easy-score/src/components/Nav.js
@@ -1,6 +1,9 @@
 import React from "react";
 import { connect } from "react-redux";
 import { Link } from "react-router-dom";
+import Hidden from "@material-ui/core/Hidden";
+import FilterListIcon from "@material-ui/icons/FilterList";
+
 import {
   setKeywordFilterValue,
   setLevelFilterValue,
@@ -8,13 +11,9 @@ import {
   setTimeFilterValue,
   setRequirementsFilterValue,
 } from "../store/actions/filtersActions";
-
 import { setShowModal } from "../store/actions/index";
 
 import SearchForm from "./SearchForm";
-import Hidden from "@material-ui/core/Hidden";
-import FilterListIcon from "@material-ui/icons/FilterList";
-
 import FiltersModal from "./FiltersModal";
 
 const Nav = (props) => {
