+++ the-easy-score/src/components/NoResults.js
@@ -1,5 +1,6 @@
 import React from "react";
 import { Empty } from "antd";
+
 const NoResults = () => {
   return (
     <div className="no-results-background-">
