+++ the-easy-score/src/components/SearchForm.js
@@ -1,6 +1,8 @@
 import React, { useEffect } from "react";
 import { useHistory } from "react-router-dom";
 import { connect } from "react-redux";
+import { FormControl, Button, InputGroup } from "react-bootstrap";
+
 import {
   setKeywordFilterValue,
   setLevelFilterValue,
@@ -9,7 +11,6 @@ import {
   setTimeFilterValue,
   resetFilterValues,
 } from "../store/actions/filtersActions";
-
 import {
   requirementsValues,
   courseLevelValues,
@@ -17,9 +18,7 @@ import {
   timeofDayValues,
 } from "../data/FilterSelectsData.js";
 
-import { FormControl, Button, InputGroup } from "react-bootstrap";
 import FilterSelect from "./FilterSelect";
-
 import "../screens/home/home.css";
 
 const SearchForm = (props) => {
