+++ the-easy-score/src/components/Sidebar.js
@@ -2,8 +2,8 @@ import React, { useState } from "react";
 import { Link } from "react-router-dom";
 import MenuIcon from "@material-ui/icons/Menu";
 import CloseIcon from "@material-ui/icons/Close";
-import { SidebarData } from "../data/SidebarData";
 
+import { SidebarData } from "../data/SidebarData";
 import "../componentsStyles/Sidebar.css";
 
 const Sidebar = () => {
