+++ the-easy-score/src/screens/coursepage/Course.js
@@ -1,6 +1,7 @@
 import React, { useEffect, useState } from "react";
 import { useParams } from "react-router-dom";
 import { connect, useDispatch } from "react-redux";
+
 import { setNavStyle, setFooterStyle } from "../../store/actions/index";
 import { fetchCoursePage } from "../../store/actions/fetchDataActions";
