+++ the-easy-score/src/screens/searchresults/components/CourseCard.js
@@ -1,6 +1,7 @@
 import React from "react";
 import { Link } from "react-router-dom";
 import ArrowForwardIcon from "@material-ui/icons/ArrowForward";
+
 import "../results.css";
 
 const CourseCard = ({ course }) => {
