+++ the-easy-score/src/screens/searchresults/components/Courses.js
@@ -1,4 +1,5 @@
 import React from "react";
+
 import CourseCard from "./CourseCard";
 
 const Courses = (props) => {
