+++ the-easy-score/src/screens/searchresults/helpers/index.js
