+++ the-easy-score/src/store/store.js
@@ -1,7 +1,6 @@
 import { createStore, applyMiddleware } from "redux";
 import { rootReducer } from "./rootReducer";
 import thunk from "redux-thunk";
-// import logger from "redux-logger";
 
 export const store = createStore(
   rootReducer, applyMiddleware(thunk)
